% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhh4ZI_hhh4.R
\name{hhh4ZI.hhh4}
\alias{hhh4ZI.hhh4}
\title{Refit a HHH4 model to a HHH4ZI model}
\usage{
\method{hhh4ZI}{hhh4}(object, zi = list(f = ~-1, lag = 1, lag.unitSpecific = FALSE))
}
\description{
Refit a HHH4 model to a HHH4ZI model
}
