% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhh4ZI.R
\name{hhh4ZI}
\alias{hhh4ZI}
\alias{hhh4ZI.sts}
\title{Fitting zero-inflated HHH Models with Random Effects
and Neighbourhood Structure}
\usage{
hhh4ZI(object, control, ...)

\method{hhh4ZI}{sts}(
  stsObj,
  control = list(ar = list(f = ~-1, offset = 1, lag = 1), ne = list(f = ~-1, offset =
    1, lag = 1, weights = neighbourhood(stsObj) == 1, scale = NULL, normalize = FALSE),
    end = list(f = ~1, offset = 1), zi = list(f = ~1, lag = 1, lag.unitSpecific = FALSE),
    family = c("NegBin1", "NegBinM"), subset = 2:nrow(stsObj), optimizer = list(stop =
    list(tol = 1e-05, niter = 100), regression = list(method = "nlminb"), variance =
    list(method = "Nelder-Mead")), verbose = FALSE, start = list(fixed = NULL, random =
    NULL, sd.corr = NULL),      data = list(t = stsObj@epoch - min(stsObj@epoch)),
    keep.terms = FALSE),
  check.analyticals = FALSE,
  ...
)
}
\arguments{
\item{control}{a list containing the model specification and control arguments,
 the parts relating to \code{hhh4} model is the same as in \code{surveillance::hhh4}:
\itemize{
\item{\code{ar}}{
Model for the autoregressive component given as
list with the following components:
\itemize{
\item{f = ~ -1}{
a formula specifying
\eqn{\log(\lambda_{it})}{log(\lambda_it)}.}
\item{offset = 1}{
optional multiplicative offset, either 1 or
a matrix of the same dimension as \code{observed(stsObj)}.}
\item{lag = 1}{
a positive integer meaning autoregression on
\eqn{y_{i,t-lag}}.}
}
}
\item{\code{ne}}{
Model for the neighbour-driven component given as
list with the following components:
 \itemize{
  \item{f = ~ -1}{
  a formula specifying \eqn{\log(\phi_{it})}{log(\phi_it)}}.
  \item{offset = 1}{
  optional multiplicative offset, either 1 or
     a matrix of the same dimension as \code{observed(stsObj)}.}
   \item{lag = 1}{
   a non-negative integer meaning dependency on
    \eqn{y_{j,t-lag}}.}
  \item{weights = neighbourhood(stsObj) == 1}{
     neighbourhood weights \eqn{w_{ji}}{w_ji}. The default
     corresponds to the original formulation by Held et al
     (2005), i.e., the spatio-temporal component incorporates an
     unweighted sum over the lagged cases of the first-order
     neighbours. See Paul et al (2008) and Meyer and Held (2014)
     for alternative specifications, e.g.,
     \code{\link{W_powerlaw}}.
     Time-varying weights are possible by specifying an
     array of \code{dim()} \code{c(nUnits, nUnits, nTime)}, where
     \code{nUnits=ncol(stsObj)} and \code{nTime=nrow(stsObj)}.}
   \item{scale = NULL}{
     optional matrix of the same dimensions as \code{weights} (or
    a vector of length \code{ncol(stsObj)}) to scale the
     \code{weights} to \code{scale * weights}.
   }
   \item{normalize = FALSE}{
    logical indicating if the (scaled) \code{weights} should be
     normalized such that each row sums to 1.
   }
 }
 }
\item{\code{end}}{
Model for the endemic component given as list
with the following components
\itemize{
\item{f = ~ 1 }{
a formula specifying \eqn{\log(\nu_{it})}{log(\nu_it)}}.
\item{offset = 1 }{
optional multiplicative offset \eqn{e_{it}}{e_it},
either 1 or a matrix of the same dimension as \code{observed(stsObj)}.}
}}
\item{\code{zi}}{
Model for the zero inflation component given as
list with the following components:
\itemize{
\item{f = ~ -1}{
a formula specifying
\eqn{logit(\gamma_{it})}{logit(gamma_it)}.}
\item{lag = 1}{
a vector of positive integers meaning autoregression on the respective
\eqn{y_{i,t-lag}} terms. Use \code{NULL} or \code{integer(0)} to exclude AR terms.}
\item{lag.unitSpecific}{ logical indicating if the autoregressive parameter
in the zero inflation part is unit specific. }
}
}
\item{\code{family}}{
Distributional family --
the Negative Binomial distribution. The
overdispersion parameter can be assumed to be the same for all
units (\code{"NegBin1"}), to vary freely over all units
(\code{"NegBinM"}), or to be shared by some units (specified by
a factor of length \code{ncol(stsObj)} such that its number of
levels determines the number of overdispersion parameters).
Note that \code{"NegBinM"} is equivalent to
\code{factor(colnames(stsObj), levels = colnames(stsObj))}.
}
\item{\code{subset}}{
Typically \code{2:nrow(obs)} if model contains
autoregression.}
\item{\code{optimizer}}{
a list of three lists of control arguments.

The \code{"stop"} list specifies two criteria for the outer
optimization of regression and variance parameters: the relative
\code{tol}erance for parameter change using the criterion
\code{max(abs(x[i+1]-x[i])) / max(abs(x[i]))},
and the maximum number \code{niter} of outer iterations.

Control arguments for the single optimizers are specified in the
lists named \code{"regression"} and \code{"variance"}.
\code{method="nlminb"} is the default optimizer for regression update,
 however, the \code{method}s from \code{\link{optim}} may also be specified
(as well as \code{"\link{nlm}"} but that one is not recommended here).
For the variance updates, only Nelder-Mead optimization
(\code{method="Nelder-Mead"}) is provided.
All other elements of these two lists are passed as
\code{control} arguments to the chosen \code{method}, e.g., if
\code{method="nlminb"} adding \code{iter.max=50} increases the
maximum number of inner iterations from 20 (default) to 50.
}

\item{\code{verbose}}{
non-negative integer (usually in the range
\code{0:3}) specifying the amount of tracing information to be
output during optimization.}

\item{\code{start}}{
a list of initial parameter values replacing
initial values set via \code{\link{fe}} and \code{\link{ri}}.
Since \pkg{surveillance} 1.8-2, named vectors are matched
against the coefficient names in the model (where unmatched
start values are silently ignored), and need not be complete,
e.g., \code{start = list(fixed = c("-log(overdisp)" = 0.5))}
(default: 2) for a \code{family = "NegBin1"} model.
In contrast, an unnamed start vector must specify the full set
of parameters as used by the model.}

\item{\code{data}}{
a named list of covariates that are to be
included as fixed effects (see \code{\link{fe}}) in any of the 3
component formulae.
By default, the time variable \code{t} is available and used for
seasonal effects created by \code{\link{addSeason2formula}}.
In general, covariates in this list can be either vectors of
length \code{nrow(stsObj)} interpreted as time-varying but
common across all units, or matrices of the same dimension as
the disease counts \code{observed(stsObj)}.}
\item{\code{keep.terms}}{ logical indicating if the terms object
used in the fit is to be kept as part of the returned object.
This is usually not necessary, since the terms object is
reconstructed by the \code{\link{terms}}-method for class
\code{"hhh4ZI"} if necessary (based on \code{stsObj} and
\code{control}, which are both part of the returned
\code{"hhh4ZI"} object).}
}}

\item{stsObj}{object of class \code{"\linkS4class{sts}"} containing the (multivariate)
count data time series.}

\item{check.analyticals}{{logical (or a subset of
\code{c("numDeriv", "maxLik")}), indicating if (how) the implemented
analytical score vector and Fisher information matrix should be
checked against numerical derivatives at the parameter starting values,
using the packages \pkg{numDeriv} and/or \pkg{maxLik}. If activated,
\code{hhh4} will return a list containing the analytical and numerical
derivatives for comparison (no ML estimation will be performed).
This is mainly intended for internal use by the package developers.}}
}
\value{
\code{hhh4ZI} returns an object of class \code{"hhh4ZI"},
which inherits from class \code{"hhh4"}, and
is a list containing the following components:
\itemize{
\item{coefficients}{
named vector with estimated (regression) parameters of the model}
\item{se}{
estimated standard errors (for regression parameters)}
\item{cov}{
covariance matrix (for regression parameters)}
\item{Sigma}{
estimated variance-covariance matrix of random effects}
\item{Sigma.orig}{
estimated variance parameters on internal scale used for optimization}
\item{call}{ the matched call }
\item{dim}{ vector with number of fixed and random effects in the model }
\item{loglikelihood}{ (penalized) loglikelihood evaluated at the MLE}
\item{margll}{ (approximate) log marginal likelihood should the model contain random effects  }
\item{convergence}{ logical. Did optimizer converge?}
\item{mu}{ The fitted mean values in hhh4 model part}
\item{fitted.values}{ fitted mean values in zero-inflated model}
\item{gamma}{ fitted zero inflation parameter}
\item{control}{ control object of the fit}
\item{terms}{ the terms object used in the fit if \code{keep.terms = TRUE}
  and \code{NULL} otherwise}
\item{stsObj}{ the supplied \code{stsObj} }
\item{lags}{ named integer vector of length three containing the lags
  used for the epidemic components \code{"ar"}, \code{"ne"} and \code{"zi"}
  respectively. The corresponding lag is \code{NA} if the component
  was not included in the model.}
\item{nObs}{ number of observations used for fitting the model}
\item{nTime}{ number of time points used for fitting the model }
\item{nUnit}{ number of units (e.g. areas) used for fitting the model}
\item{runtime}{ the \code{\link{proc.time}}-queried time taken
  to fit the model, i.e., a named numeric vector of length 5 of class
  \code{"proc_time"}}
}
}
\description{
Fits a zero-inflated autoregressive negative binomial (\code{hhh4}) model
to a univariate or multivariate time series of counts.
The characteristic feature of \code{hhh4} models is the additive
decomposition of the conditional mean into \emph{epidemic} and
\emph{endemic} components (Held et al, 2005).
The inflated parameter is a logit-linear predictor and can have autoregressive terms.
}
\examples{
adjmat <- neighbourhood(measles) == 1
neW1 <- adjmat/colSums(adjmat)  # FIXME @ Junyi: why normalize columns?
fit <- hhh4ZI(measles,
  control = list(
    ar = list(f = ~1),
    ne = list(f = ~1, weights = neW1, normalize = TRUE),
   end = list(f = ~1),
    zi = list(f = ~1),
    family = "NegBin1",
    verbose = TRUE,
    keep.terms = TRUE
  )
)
summary(fit)
sim_data <- simulate(fit, simplify = FALSE)
}
