% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\docType{data}
\name{measles}
\alias{measles}
\title{Measles in the 16 states of Germany}
\format{
An object of class \code{sts} with 364 rows and 16 columns.
}
\source{
SurvStat@RKI 2.0 (https://survstat.rki.de),
Robert Koch Institute; Queried on 26 March 2021.

Statistisches Bundesamt (https://www.destatis.de/DE/Home/_inhalt.html);
Queried on 10 April 2021.
}
\usage{
measles
}
\description{
Bi-weekly number of measles cases in the 16 states (Bundeslaender) of Germany
for the years 2005 to 2018.
}
\details{
The \code{population} slot contains the population fractions
of each state from 2005 to 2018, obtained from the Federal Statistical Office of Germany.
}
\keyword{datasets}
