% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\docType{data}
\name{measlesDE}
\alias{measlesDE}
\title{Measles in the 16 states of Germany}
\format{
An \code{sts} object containing \eqn{749\times 16}{749 x 16} observations starting
from week 1 in 2005.The \code{population} slot contains the population fractions
of each state at 31.12.2006, obtained from the Federal Statistical Office of Germany.
}
\source{
SurvStat@RKI 2.0 (https://survstat.rki.de),
Robert Koch Institute; Queried on 26 March 2021.
}
\usage{
data(measlesDE)
}
\description{
Weekly number of measles cases in the 16 states (Bundeslaender) of Germany
for years 2005 to 2019.
}
\keyword{datasets}
