% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhh4ZI_onestepahead.R
\name{oneStepAhead}
\alias{oneStepAhead}
\alias{oneStepAhead.hhh4ZI}
\title{Predictive Model Assessment for hhh4ZI Models}
\usage{
oneStepAhead(result, tp, ...)

\method{oneStepAhead}{hhh4ZI}(
  result,
  tp,
  type = c("rolling", "first", "final"),
  which.start = c("current", "final"),
  keep.estimates = FALSE,
  verbose = TRUE,
  cores = 1
)
}
\description{
The function oneStepAhead computes successive one-step-ahead
 predictions for a (random effects) HHH4ZI model fitted by \code{hhh4ZI.}
This function is the equivalent of \code{surveillance::oneStepAhead}
for model fits of class
\code{hhh4ZI}, obtained from \code{hhh4ZI}. The arguments are the
same as in \code{surveillance::oneStepAhead}.
}
