% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pit.R
\name{pit.oneStepAhead_hhh4ZI}
\alias{pit.oneStepAhead_hhh4ZI}
\alias{pit.hhh4ZI}
\title{Non-Randomized Version of the PIT Histogram (for Count Data)}
\usage{
\method{pit}{oneStepAhead_hhh4ZI}(x, units = NULL, ...)

\method{pit}{hhh4ZI}(x, subset = x$control$subset, units = seq_len(x$nUnit), ...)
}
\description{
Non-Randomized Version of the PIT Histogram (for Count Data)
}
