% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simHHH4zi.R
\name{simulate.hhh4ZI}
\alias{simulate.hhh4ZI}
\title{Simulate "hhh4ZI" Count Time Series}
\usage{
\method{simulate}{hhh4ZI}(
  object,
  nsim = 1,
  seed = NULL,
  y.start = NULL,
  subset = 1:nrow(object$stsObj),
  coefs = coef(object),
  components = c("ar", "ne", "end"),
  simplify = nsim > 1,
  ...
)
}
\description{
This function is the equivalent of \code{surveillance::simulate.hhh4} for model fits of class
\code{hhh4ZI}, obtained from \code{hhh4ZI}. The arguments are the
same as in \code{surveillance::simulate.hhh4}.
}
