% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhh4ZI_methods.R
\name{update.hhh4ZI}
\alias{update.hhh4ZI}
\title{update a fitted "hhh4ZI" model}
\usage{
\method{update}{hhh4ZI}(
  object,
  ...,
  S = NULL,
  subset.upper = NULL,
  use.estimates = object$convergence,
  evaluate = TRUE
)
}
\description{
Re-fit a \code{hhh4ZI} model with a modified control list.
This function is the equivalent of \code{surveillance::update.hhh4}
for model fits of class
\code{hhh4ZI}, obtained from \code{hhh4ZI}. The arguments are the
same as in \code{surveillance::update.hhh4}.
}
