% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhh4ZI_hhh4.R
\name{upgrade.hhh4}
\alias{upgrade.hhh4}
\title{Refit a HHH4 model with a ZI Component}
\usage{
\method{upgrade}{hhh4}(object, control = list(f = ~1, lag = 1, lag.unitSpecific = FALSE), ...)
}
\arguments{
\item{object}{a \code{"\link{hhh4}"} fit.}

\item{control}{either a list of specifications for the \code{zi} component
of \code{\link{hhh4ZI}} or a \code{hhh4} control list with added \code{zi}
specification.}

\item{...}{further (non-\code{zi}) control elements.}
}
\description{
Refits a previous \code{\link{hhh4}} fit with a ZI component
using \code{\link{hhh4ZI}}.
}
