% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pit.R
\name{pit.default}
\alias{pit.default}
\alias{plot.pit}
\alias{pit.oneStepAhead_hhh4ZI}
\alias{pit.hhh4ZI}
\title{Non-Randomized Version of the PIT Histogram (for Count Data)}
\usage{
\method{pit}{default}(x, pdistr, J = 10, relative = TRUE, ..., plot = list())

\method{plot}{pit}(x, main = "", ylab = NULL, ...)

\method{pit}{oneStepAhead_hhh4ZI}(x, units = NULL, ...)

\method{pit}{hhh4ZI}(x, subset = x$control$subset, units = seq_len(x$nUnit), ...)
}
\description{
Non-Randomized Version of the PIT Histogram (for Count Data)
}
