% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhh4ZI_plot.R
\name{plot.hhh4ZI}
\alias{plot.hhh4ZI}
\alias{plotHHH4ZI_fitted}
\alias{plotHHH4ZI_fitted1}
\alias{plotHHH4ZI_maps}
\alias{plotHHH4ZI_ri}
\alias{plotHHH4ZI_season}
\alias{plotHHH4ZI_neweights}
\title{lots for Fitted hhh4ZI-models}
\usage{
plot.hhh4ZI(x, type = c("fitted", "season", "maps", "ri", "neweights"), ...)

plotHHH4ZI_fitted(
  x,
  units = 1,
  names = NULL,
  col = c("grey85", "blue", "orange"),
  pch = 19,
  pt.cex = 0.6,
  pt.col = 1,
  par.settings = list(),
  legend = TRUE,
  legend.args = list(),
  legend.observed = FALSE,
  decompose = NULL,
  total = FALSE,
  meanHHH = NULL,
  ...
)

plotHHH4ZI_fitted1(
  x,
  unit = 1,
  main = NULL,
  col = c("grey85", "blue", "orange"),
  pch = 19,
  pt.cex = 0.6,
  pt.col = 1,
  border = col,
  start = x$stsObj@start,
  end = NULL,
  xaxis = NULL,
  xlim = NULL,
  ylim = NULL,
  xlab = "",
  ylab = "No. infected",
  hide0s = FALSE,
  decompose = NULL,
  total = FALSE,
  meanHHH = NULL
)

plotHHH4ZI_maps(
  x,
  which = c("mean", "endemic", "epi.own", "epi.neighbours", "gamma"),
  prop = FALSE,
  main = which,
  zmax = NULL,
  col.regions = NULL,
  labels = FALSE,
  sp.layout = NULL,
  ...,
  map = x$stsObj@map,
  meanHHH = NULL
)

plotHHH4ZI_ri(
  x,
  component,
  exp = FALSE,
  at = list(n = 10),
  col.regions = cm.colors(100),
  colorkey = TRUE,
  labels = FALSE,
  sp.layout = NULL,
  gpar.missing = list(col = "darkgrey", lty = 2, lwd = 2),
  ...
)

plotHHH4ZI_season(
  ...,
  components = NULL,
  intercept = FALSE,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = "",
  main = NULL,
  par.settings = list(),
  matplot.args = list(),
  legend = NULL,
  legend.args = list(),
  refline.args = list(),
  unit = 1
)

plotHHH4ZI_neweights(x, plotter = boxplot, ..., exclude = 0, maxlag = Inf)
}
\description{
This function is the equivalent of \code{surveillance::plot.hhh4} for model fits of class
\code{hhh4ZI}, obtained from \code{hhh4ZI}. The arguments are the
same as in \code{surveillance::plot.hhh4}, with one more component option for zero inflation part..
}
