% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{scores}
\alias{scores}
\alias{scores.oneStepAhead_hhh4ZI}
\alias{scores.hhh4ZI}
\title{Proper Scoring Rules for Poisson, Negative Binomial
or Zero-inflated Negative Binomial Predictions}
\usage{
\method{scores}{oneStepAhead_hhh4ZI}(
  x,
  which = c("logs", "rps", "dss", "ses"),
  units = NULL,
  sign = FALSE,
  individual = FALSE,
  reverse = FALSE,
  ...
)

\method{scores}{hhh4ZI}(
  x,
  which = c("logs", "rps", "dss", "ses"),
  subset = x$control$subset,
  units = seq_len(x$nUnit),
  sign = FALSE,
  ...
)
}
\description{
The following scores are implemented:
logarithmic score (logs), ranked probability score (rps),
Dawid-Sebastiani score (dss), squared error score (ses).
These function are equivalent of \code{surveillance::scores},
\code{surveillance::logs}, \code{surveillance::rps},
\code{surveillance::dss}, \code{surveillance::ses},
further extened for zero-inflated negative binomial predictions.
The arguments for Possion and negative binomial distribution predictions
are the same as in \code{surveillance::update.hhh4},
for zero-inflated negative binomial predictions the zero inflation parameter
\code{gamma} is needed.
}
