% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{logs}
\alias{logs}
\alias{ses}
\alias{nses}
\alias{dss}
\alias{rps}
\alias{scores}
\alias{scores.default}
\alias{scores.oneStepAhead_hhh4ZI}
\alias{scores.hhh4ZI}
\title{Proper Scoring Rules for Poisson, Negative Binomial
or Zero-inflated Negative Binomial Predictions.}
\usage{
logs(x, mu, size = NULL, gamma = NULL)

ses(x, mu, size = NULL, gamma = NULL)

nses(x, mu, size = NULL, gamma = NULL)

dss(x, mu, size = NULL, gamma = NULL)

rps(
  x,
  mu,
  size = NULL,
  gamma = NULL,
  k = 40,
  tolerance = sqrt(.Machine$double.eps)
)

scores(result, tp, ...)

\method{scores}{default}(
  x,
  mu,
  size = NULL,
  gamma = NULL,
  which = c("logs", "rps", "dss", "ses"),
  sign = FALSE,
  ...
)

\method{scores}{oneStepAhead_hhh4ZI}(
  x,
  which = c("logs", "rps", "dss", "ses"),
  units = NULL,
  sign = FALSE,
  individual = FALSE,
  reverse = FALSE,
  ...
)

\method{scores}{hhh4ZI}(
  x,
  which = c("logs", "rps", "dss", "ses"),
  subset = x$control$subset,
  units = seq_len(x$nUnit),
  sign = FALSE,
  ...
)
}
\description{
The following scores are implemented:
logarithmic score (logs), ranked probability score (rps),
Dawid-Sebastiani score (dss), squared error score (ses).
These function are equivalent of \code{surveillance::scores},
\code{surveillance::logs}, \code{surveillance::rps},
\code{surveillance::dss}, \code{surveillance::ses},
further extened for zero-inflated negative binomial predictions.
The arguments for Possion and negative binomial distribution predictions
are the same as in \code{surveillance::update.hhh4},
for zero-inflated negative binomial predictions the zero inflation parameter
\code{gamma} is needed.
}
